pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-camera-ocr"
version = "0.1.0"
description = "OCR from virtual camera every N seconds using Gemini API"
requires-python = ">=3.10"
dependencies = [
  "opencv-python>=4.8",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "google-genai>=0.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
